feat: separate tests and mocks
