--- src/App.test.tsx
@@ -1,14 +1,0 @@
-import { describe, it } from 'vitest';
-import { render, screen } from '@testing-library/react';
-import { App } from './App';
-
-describe('App', () => {
-  it('Test render home page', () => {
-    render(<App />);
-    expect(
-      screen.getByRole('heading', {
-        level: 1,
-      })
-    ).toHaveTextContent('Home');
-  });
-});
