+++ src/__mocks__/response.mock.ts
@@ -1,0 +1,30 @@
+const mockResponse = {
+  products: [
+    {
+      category: 'book',
+      description: 'The Boy Who Lived Has Come To Die',
+      id: 999,
+      image: 'http://url.com',
+      price: 8888,
+      rating: {
+        count: 77777,
+        rate: 4.9,
+      },
+      title: 'Harry Potter',
+    },
+    {
+      category: 'kitchen',
+      description: 'Big size',
+      id: 123,
+      image: 'http://url2.com',
+      price: 1234,
+      rating: {
+        count: 12345,
+        rate: 0.1,
+      },
+      title: 'Silver plate',
+    },
+  ],
+};
+
+export { mockResponse };
