+++ src/components/CustomForm/ErrorMessage/__tests__/ErrorMessage.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { ErrorMessage } from './ErrorMessage';
+import { ErrorMessage } from '../ErrorMessage';
 
 describe('ErrorMessage', () => {
   it('Test render error with message', () => {
