+++ src/components/CustomForm/StatusMessage/__tests__/StatusMessage.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { StatusMessage } from './StatusMessage';
+import { StatusMessage } from '../StatusMessage';
 
 describe('StatusMessage', () => {
   it('Test render status', () => {
