+++ src/components/RatingLogo/__tests__/RatingLogo.test.tsx
@@ -1,10 +1,10 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { StockLogo } from './StockLogo';
+import { RatingLogo } from '../RatingLogo';
 
-describe('StockLogo', () => {
-  it('Test render stock logo', () => {
-    render(<StockLogo />);
+describe('RatingLogo', () => {
+  it('Test render rating logo', () => {
+    render(<RatingLogo />);
     expect(screen.getByRole('img')).toBeInTheDocument();
   });
 });
