+++ src/pages/__tests__/Forms.test.tsx
@@ -1,14 +1,14 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { NotFound } from './NotFound';
+import { Forms } from '../Forms';
 
-describe('NotFound', () => {
+describe('Forms', () => {
   it('Test render NotFound', () => {
-    render(<NotFound />);
+    render(<Forms />);
     expect(
       screen.getByRole('heading', {
         level: 1,
       })
-    ).toHaveTextContent('404');
+    ).toHaveTextContent('Forms');
   });
 });
