+++ src/utils/__mocks__/stringHelpers.mock.ts
@@ -1,0 +1,4 @@
+const mock = 'asd';
+const result = 'Asd';
+
+export { mock, result };
