+++ src/utils/__tests__/stringHelpers.test.tsx
@@ -1,12 +1,10 @@
 import { describe, it } from 'vitest';
-import { capitalize } from './stringHelpers';
-
-const mock = 'asd';
-const expectResult = 'Asd';
+import { capitalize } from '../stringHelpers';
+import { mock, result } from '../__mocks__/stringHelpers.mock';
 
 describe('String helpers', () => {
   it('Test capitalize strings', () => {
-    expect(capitalize(mock)).toEqual(expectResult);
+    expect(capitalize(mock)).toEqual(result);
   });
   it('Test capitalize strings', () => {
     expect(capitalize()).toEqual('');
